refactor(coderd/x/chatd): move OpenAI computer use helpers
